requirements.txt
lightning
matplotlib
torch
torchaudio
torchmetrics
torchvision
hydra-core
wandb
diffusers
transformers
rasterio
einops